Got multiple test files working with mocha-loader
