+++ webpack.config.js
@@ -10,7 +10,7 @@ module.exports = {
             { test: /\.js$/, exclude: /node_modules/, loader: "eslint-loader" }
         ],
         loaders: [
-            { test: /\.js$/, exclude: /node_modules/, loader: "babel-loader" }
+            { test: /\.js$/, exclude: /node_modules/, loader: "babel-loader" },
         ]
     }
 };
